data_repository.gd needed  changes to deal with a local user, a signal is emited on line 84 to update the board
so cards switching lists can be updated and saved locally.
Some methods were rearranged to navigate easier methods that affect the same object eg. boards lists and cards.
An if statement was needed in the update_card method to deal with loading cards as resources that would
trigger an error because as first items to be loaded they would find an existing list or board.
Added a method _on_list_dropped that is planned to deal with lists being swapped in the interface, atm they are only
visually working but there is no position data available yet.